Not A Compression Submission
To get things started here is something that can be used for further improvement.

It is not very fast. It is as efficient as Snappy/LZ4 and uses a similar approach.

Original length: 3295450, Encoded length: 1527541, Compression ratio: 0.46

SUBMITTERS ARE FREE TO USE THIS CODE! (unless organizers disagree)

Use it as a baseline and put your improvements on top.
The most complicated concept we use is uvarint.

I have noted the most obvious improvements, but there are many more.